'profile_set

+++ app/Http/Controllers/NutritionalItemController.php
@@ -5,11 +5,13 @@ namespace App\Http\Controllers;
 use App\Http\Requests\NutritionalItemRequest;
 use Illuminate\Http\Request;
 use App\Models\NutritionalItem;
+use Illuminate\Support\Facades\Gate;
 
 class NutritionalItemController extends Controller
 {
     public function index()
     {
+        Gate::authorize('browse_nutritional_items');
         $dataType = 'nutritionalItem';
         $data = NutritionalItem::all();
 
@@ -21,6 +23,7 @@ class NutritionalItemController extends Controller
 
     public function store(Request $request)
     {
+        Gate::authorize('create_nutritional_items');
         $nutritionalItem = NutritionalItem::create($request->all());
 
         return redirect()->route('nutritionalitems.index')->with('message', 'Created Successfully!');
@@ -28,6 +31,7 @@ class NutritionalItemController extends Controller
 
     public function show(NutritionalItem $nutritionalItem)
     {
+        Gate::authorize('show_nutritional_items');
         $dataType = 'nutritionalItem';
         $data = $nutritionalItem;
 
@@ -40,6 +44,7 @@ class NutritionalItemController extends Controller
     
     public function edit(NutritionalItem $nutritionalItem)
     {
+        Gate::authorize('update_nutritional_items');
         $dataType = 'nutritionalItem';
         $data = $nutritionalItem;
 
@@ -51,6 +56,7 @@ class NutritionalItemController extends Controller
 
     public function create()
     {
+        Gate::authorize('create_nutritional_items');
         $dataType = 'nutritionalItem';
         $data = null;
 
@@ -63,6 +69,8 @@ class NutritionalItemController extends Controller
     public function update(Request $request,NutritionalItem $nutritionalItem)
     {
 
+        Gate::authorize('update_nutritional_items');
+
         $nutritionalItem->update($request->all());
 
         return redirect()->route('nutritionalitems.index')->with('message', 'Updated Successfully!');
@@ -70,6 +78,7 @@ class NutritionalItemController extends Controller
 
     public function destroy(NutritionalItem $nutritionalItem)
     {
+        Gate::authorize('delete_nutritional_items');
         $nutritionalItem->delete();
 
         return redirect()->route('nutritionalitems.index')->with('message', 'Deleted Successfully!');
@@ -78,7 +87,7 @@ class NutritionalItemController extends Controller
 
 
     public function bulkDestroy(Request $request){
-
+        Gate::authorize('delete_nutritional_items');
         NutritionalItem::destroy($request->ids);
         return redirect()->route('nutritionalitems.index')->with('message', 'Deleted Successfully!');
     }
